replace currency logic in currency folder and fix name file
