+++ src/redux/auth/auth-reducer.js
@@ -20,7 +20,6 @@ const error = createReducer(null, {
   [authActions.loginError]: setError,
   [authActions.statisticsError]: setError,
   [authActions.fetchStatisticsError]: setError,
-  [authActions.currencyError]: setError,
   [authActions.getTokenFromLSError]: setError,
 });
 
@@ -34,21 +33,9 @@ const isAuthenticated = createReducer(false, {
   [authActions.logoutSuccess]: () => false,
 });
 
-const currency = createReducer(null, {
-  [authActions.currencySuccess]: (_, { payload }) => payload.data,
-});
-
-const loading = createReducer(false, {
-  [authActions.currencyRequest]: () => true,
-  [authActions.currencySuccess]: () => false,
-  [authActions.currencyError]: () => false,
-});
-
 export default combineReducers({
   isAuthenticated,
   user,
   token,
   error,
-  currency,
-  loading,
 });
