+++ src/redux/currency/currency-selectors.js
@@ -1,0 +1,2 @@
+const getCurrency = state => state.currency.data;
+export default { getCurrency };
