+++ src/redux/currency/index.js
@@ -1,0 +1,4 @@
+export { default as currencyActions } from './currency-actions';
+export { default as currencyOperations } from './currency-operations';
+export { default as currencyReducer } from './currency-reducer';
+export { default as currencySelectors } from './currency-selectors';
